blog/requirements.txt
dj-database-url==0.4.2
Django==1.10.7
gunicorn==19.7.1
olefile==0.44
Pillow==4.1.1
whitenoise==3.3.0
django-ckeditor==5.2.2
django-el-pagination==3.1.0
django-mptt==0.8.7
django-static-sitemaps==4.2.1
psycopg2==2.7.1
